Reorganize listing related files in the widgets direct
Move listing logic to own package, create separate file for column classes.

WIDGETS FRNT-2 [IN PROGRESS]
Table FRNT-3 [IN PROGRESS]
