+++ BaCa2/main/views.py
@@ -12,7 +12,8 @@ from util.models import model_cls
 from util.views import normalize_string_to_python
 from main.models import (Course, User)
 from widgets.base import Widget
-from widgets.listing import (TableWidget, TextColumn)
+from widgets.listing import TableWidget
+from widgets.listing.columns import TextColumn
 from widgets.forms import FormWidget
 from widgets.forms.fields import get_field_validation_status
 from widgets.navigation import (NavBar, SideNav)
