+++ BaCa2/widgets/listing/__init__.py
@@ -1,0 +1,1 @@
+from .base import TableWidget, TableWidgetPaging
